Renamed OrderingApp as OrderApp
